// features/qlik-app-info/types.ts
export type AppPropsResponseType = {
  id: string;
  modifiedDate: string;
  owner: string;
  publishTime: string;
  published: boolean;
  qLastReloadTime: string;
  qSavedInProductVersion: string;
  qTitle: string;
  qUsage: string;
  spaceId: string;
};

export type AppPropsType = {
  id: string;
  modifiedDate: string;
  owner: string;
  publishTime: string;
  published: string;
  lastReloadTime: string;
  version: string;
  title: string;
  usage: string;
  spaceId: string;
};